requirements.txt
aiohttp==3.9.1
asyncio-throttle==1.0.2
APScheduler==3.10.4
requests==2.31.0
python-dotenv==1.0.0
psutil==5.9.6
reportlab==4.0.7
lxml==4.9.3
python-dateutil==2.8.2
pytz==2023.3
aiosqlite==0.19.0
matplotlib==3.8.2
pandas==2.1.4
numpy==1.26.2
email-validator==2.1.0
pydantic==2.5.2
aiofiles==24.1.0